refactor: update file structures
